Add markdown support for the security patch details fields
It turns out that a lot of the detailed descriptions can benefit
from things like paragraphs and some formatting, so it is better
to render it in markdown.

This does not show the markdown in a preview view due to a bunch
of reasons, but future work can improve upon that.