Fix/remove userse of dictsort function
As of django 3.2.11 in what's considered a security fix (wouldn't be in
our use cases, but could be in others), we can no longer use dictsort
on lower() -- doing so returns an empty list.

So instead we create a new filter called sort_lower, and use that one
fromt he two places where we use the filter.

This bug caused the list of translations and yearly reports no the CoC
page to become empty. Spotted by Umair Shahid.