Give priority to the "new style" of news links over the old
This affects how we redirect news when tehre's what looks like an "id"
number both before and after. For example, the link:

2016-08-11-security-update-release-1688

would previously detect the id as 2017 and redirect to that article,
which is obviously wrong.

This changes the order so that id-at-the-end is checked first. This
instead gives problems for urls that *end* in a year (or other things
that looks like an id).

This is not ideal, but it's better than before because at least now the
links that are being generated *now* are handled the correct way.