Fix display issue with form button renderings
This was likely introduced with 5389adf8 and dark mode, as
the "!important" on all input elements was overriding the
desired styles for the submission buttons in light mode.
But forcing that style is really not that important, so
we can let CSS do it's thing to render the button in its
desired state.

Reported-by: Andreas 'ads' Scherbaum <ads@pgug.de>