Allow deletion of draft news articles
When somebody posts a news article, make it possible to delete it before
it's submitted to moderation (or after it's been withdrawn or bounced),
instead of forcing the user to leave it around ForEver (TM).

Do this by adding some generic functionality for confirmation popups,
that can also be used for other things in the future.