Update community Slack links again to fix invite procedure
eccfb71c did not correctly invite people who were not in the
allow-listed domains. This uses a "shared join" link that
allows this to occur.

This is considered a temporary fix to ensure new users can still
sign up for the community Slack account. A future commit will
direct people to a better solution.

Reported-by: Stephen Frost <sfrost@snowman.net>