Refactor release notes rendering
Move to use core_version for more information, making some purges easier
and more predictable as well. Just create a hardcoded list of versions
prior to what we have real version entries for, it's just a couple that
we need to reference old release notes for (pre-6.3), so it simplifies a
lot.

Fixes (again) the release notes purging to use xkeys as set, which was
accidentally and partially undone in ac618d1b.

Generally makes the code easier to read and fixes a few cornercases.

Templates also much simplified, with the list template accidentally
already committed as part of 19682de8.

Reviewed-by: Jonathan Katz