List Fastly as CDN sponsor.
This is not the prettiest at the moment as the server sponsor page doesn't
include logos, but we are obliged to display the Fastly one in return for the
services they provide. We may want to redesign this page at some point in
the future to group by sponsor and work more cleanly with non-server hosting
sponsors.