Give better error message when invalid data is passed to cauth in d param
Int he sample cauth provider for django, the incoming data in the 'd'
parameter is supposed to be the three parts required to do a SIV
decryption of it. But if somebody is sitting on an old link or ends up
going through history or something like that it can end up with just two
parameters since that's what the old version of the plugin uses. Instead
of crashing on that, give an error message so the user can just retry.