Make docspage inherit from base.html
There was 99+% overlap between these, leading to some details missed
between them. The fact that the whole base is reproduced inside
docspage.html is probably legacy from a long time ago. So
instead make docspage.html inherit from base and set the required parametrs.

The only output difference in this is that the rel=canonical link moves
to a different part in the <head>, but it's still in <head>.

As a bonus, this also fixes the gitrev tag for theme.js, that commit 97942bcab
only fixed for base and not docs, and it fixes the footer link to
policies that e7099c90e only updated for base and not docs.